+++ src/components/NewFieldset/Fieldset.tsx
@@ -1,7 +1,0 @@
-import React, { ComponentProps } from 'react'
-
-import { FormGroup } from '../FormGroup'
-
-export const Fieldset: React.FC<ComponentProps<typeof FormGroup>> = (props) => (
-  <FormGroup {...props} as="fieldset" />
-)