+++ src/components/SectioningContent/index.ts
@@ -1,2 +1,2 @@
-export { Article, Aside, Nav, Section, SectioningFragment } from './SectioningContent'
+export { Article, Aside, Nav, Section } from './SectioningContent'
 export { LevelContext } from './levelContext'